fix(web): move formatLabel to shared module to fix case detail 500 errors (#643)
The case detail page (cases/[id]/page.tsx) imported formatLabel from
CaseDetail.tsx, a 'use client' module. In Next.js, non-component exports
from client modules are not available in server components during SSR,
causing an unhandled TypeError that resulted in 500 errors on every case
detail page.

Moved formatLabel and its helper constants (LABEL_MAP, UPPERCASE_LABEL_WORDS)
from CaseDetail.tsx to the shared lib/display-helpers.ts module, which is
not marked 'use client' and is safe to import from both server and client
components. Updated all imports and tests accordingly.
