+++ packages/web/tests/case-detail.test.ts
@@ -1,12 +1,12 @@
 import { describe, it, expect } from 'vitest';
 import {
-  formatLabel,
   truncateText,
   groupParties,
   buildDownloadUrl,
   FORMAT_LABELS,
   RULING_TEXT_TRUNCATE_LENGTH,
 } from '../src/app/cases/[id]/CaseDetail';
+import { formatLabel } from '../src/lib/display-helpers';
 
 describe('formatLabel', () => {
   it('returns em-dash for null', () => {
